Add security api request

+++ ApiBundle/Controller/GlobalController.php
@@ -292,6 +292,10 @@ abstract class GlobalController extends ApiController implements GlobalControlle
             return $this->serializeResponse("geoks.entity.notFound", Response::HTTP_NOT_FOUND);
         }
 
+        if (!in_array('ROLE_ADMIN', $this->getUser()->getRoles())) {
+            return $this->serializeResponse("geoks.user.forbidden", Response::HTTP_FORBIDDEN);
+        }
+
         $em->remove($entity);
         $em->flush();
